Added more onboarding features and fixed the calendar
The calendar was starting on the wrong date of the week so I fixed it. I also added a dark mode and streamlined the onboarding process a bit. I added a log out feature as well as a password (pin) so the switch between profiles is better and more private.